Refac: rename transform to augmentation
