+++ src/careamics/compat/transforms/n2v_manipulate.py
@@ -6,7 +6,7 @@ import numpy as np
 from numpy.typing import NDArray
 
 from careamics.config.support import SupportedPixelManipulation, SupportedStructAxis
-from careamics.dataset.transforms.transform import Transform
+from careamics.dataset.augmentation.transform import Transform
 from careamics.lightning.lightning_modules.n2v_utils.struct_mask_parameters import (
     StructMaskParameters,
 )
