+++ tests/compat/transforms/test_tta.py
@@ -3,7 +3,7 @@ import pytest
 import torch
 
 from careamics.compat.transforms.tta import ImageRestorationTTA
-from careamics.dataset.transforms import XYFlip, XYRandomRotate90
+from careamics.dataset.augmentation import XYFlip, XYRandomRotate90
 
 
 @pytest.mark.parametrize(
