+++ tests/config/augmentations/test_xy_random_rotate90_config.py
@@ -1,5 +1,5 @@
 from careamics.config.augmentations import XYRandomRotate90Config
-from careamics.dataset.transforms import XYRandomRotate90
+from careamics.dataset.augmentation import XYRandomRotate90
 
 
 def test_comptatibility_with_transform():
